requirements.txt
psycopg2-binary==2.9.9
qrcode==7.4.2
flask==2.3.3
opencv-python==4.8.1.78
scikit-learn==1.3.2